Update index.js
Work on removing any comments that aren't explicitly helpful for the reviewer. General rule of thumb is that your code should be so readable that you don't need comments to explain what it's doing. For instance, I assume that the function called `randomNumber()` generates a random number - the comment above is unnecessary. It's fine to put comments ALL OVER the place when you're in development, but make sure to remove them before submitting for evals.